Fixed bug where go-append would crash the kernel
If go_append had to grow the underlying array, the runtime
would call runtime_roundupsize. Unfortunately, the tables used
by runtime were not initialized, and everything would round up
to 0. As a result, the kernel would relocate the slice to
memory address 0, and overwrite the video buffer.

This directly mallocs the appropriate amount of memory, instead
of trying to roundup to an aligned size to avoid the issue. A better
solution would probably be to initialize the appropriate tables in
msize.c, but the code that does the initialization depends on some
symbols which are not yet defined in libg and this workaround works.